backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.24
sqlalchemy>=2.0
alembic>=1.12
pydantic>=2.5
pydantic-settings>=2.0
python-dotenv>=1.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
apscheduler>=3.10
